refactor: option 책임 완화를 위해 .json() 등을 api안으로 옮김
