+++ src/features/notifications/api/index.ts
@@ -1,2 +1,1 @@
-export { notificationRepository } from './notification.repository';
 export { notificationApi } from './notifications.api';
