--- src/features/notifications/model/notification.service.tsx
@@ -1,8 +1,0 @@
-export {
-  getNotificationDescription,
-  getNotificationViewModel,
-  notificationThumbnailKeyForType,
-  notificationTitleForType,
-} from './notification-view.utils';
-export { useNotificationReadActions } from './use-notification-read-actions';
-export { useNotificationService } from './use-notification-service';
